Ensure interface with primary IP is not deleted from remote
In cases where the primary IP used to manage a device is specified
in the inventory SoR on an interface that does not actually exist on the
infrastructure, this change ensures the non-existent interface is not deleted.
This is provisioned to ensure the inventory SoR can be populated quickly,
without need for a complete representation of DCIM and IPAM, and that
population of the complete DCIM/IPAM dataset will not overwrite the inventory
SoT's ability to manage devices.